+++ javascript/README.md
@@ -1,0 +1,18 @@
+# Practical Parser: Javascript
+
+This repository contains a simple Javascript parser which you can utilize to read input files from standard input using the Node process.
+
+A boilerplate Javascript file (`index.js`) has been provided for you to work with.  The boilerplate already includes the Javascript parser which will parse the events from an input file.
+
+By default, the boilerplate code will read and parse the file specified in the `solve` npm script.
+
+Eg: `node index.js test.txt` will read and parse the `test.txt` file in the current directory.
+
+## Steps to Run Solution
+
+1. Run `npm run solve` to run the solution
+2. Update the `solve` npm script to provide different input text files
+
+## Tests
+
+Feel free to run unit tests on the event parser via `npm test` to see the expected output from the parser
