+++ javascript/__tests__/parser.test.js
@@ -1,11 +1,12 @@
 const test = require('tape');
+const path = require('path');
 
-const parser = require('./parser.js');
+const parser = require('../parser.js');
 
 test('one event', function (t) {
   t.plan(1);
   t.deepEqual(
-    parser('1, 2, 3, 4'),
+    parser(path.join(__dirname, 'one_event.txt')),
     [{
       eventDay: 1,
       numSmall: 2,
@@ -18,7 +19,7 @@ test('one event', function (t) {
 test('multiple events', function(t) {
   t.plan(1);
   t.deepEqual(
-    parser('1, 2, 3, 4\n5, 6, 7, 8'),
+    parser(path.join(__dirname, 'multiple_events.txt')),
     [{
       eventDay: 1,
       numSmall: 2,
