+++ javascript/parser.js
@@ -1,36 +1,39 @@
 const fs = require('fs');
 
 /**
- * Takes in input string containing rows of events OR reads file from
- * standard input and formats output to an array of event objects.
+ * Reads file from standard input and formats output to an array of event objects.
  *
  * Example Input:
  *    1, 2, 4, 5
  *    6, 7, 9, 9
  *
- * Output:
- *    {
- *      eventDay: Int,
- *      numSmall: Int,
- *      numMedium: Int,
- *      numLarge: Int,
- *    }
+ * Expected Output:
+ *    [{
+ *      eventDay: 1,
+ *      numSmall: 2,
+ *      numMedium: 4,
+ *      numLarge: 5
+ *    }, {
+ *      eventDay: 6,
+ *      numSmall: 7,
+ *      numMedium: 9,
+ *      numLarge: 9,
+ *    }]
  *
- * @param {string} input String input containing comma separated event details.
+ * @param {string} filePath Path of file to events text file
  * @returns {object[]} Array containing list of events
  */
-module.exports = function parseInputToEvents(input) {
-  const inputString = input || fs.readFileSync(0).toString();
-  return inputString
+module.exports = function parseInputToEvents(filePath) {
+  return fs.readFileSync(filePath, 'utf-8')
     .split('\n')
     .filter(input => input.length > 0)
     .map(eventStr => {
-      const params = eventStr.split(',');
+      const [eventDay, numSmall, numMedium, numLarge] = eventStr.split(',');
       return {
-        eventDay: +params[0],
-        numSmall: +params[1],
-        numMedium: +params[2],
-        numLarge: +params[3],
+        eventDay: +eventDay,
+        numSmall: +numSmall,
+        numMedium: +numMedium,
+        numLarge: +numLarge,
       };
     });
 }
