+++ javascript/sample_inputs/test.txt
@@ -1,2 +1,1 @@
 5, 0, 1, 2
-6, 1, 2, 3
